Algoritime/requirements.txt
pillow 
django-admin-interface
mysqlclient